+++ r2-navigator-swift/EPUB/EPUBSpreadView.swift
@@ -1,18 +1,12 @@
 //
-//  EPUBSpreadView.swift
-//  r2-navigator-swift
-//
-//  Created by Winnie Quinn, Alexandre Camilleri, Mickaël Menu on 8/23/17.
-//
 //  Copyright 2019 Readium Foundation. All rights reserved.
-//  Use of this source code is governed by a BSD-style license which is detailed
-//  in the LICENSE file present in the project repository where this source code is maintained.
+//  Use of this source code is governed by the BSD-style license
+//  available in the top-level LICENSE file of the project.
 //
 
 import WebKit
 import R2Shared
 import SwiftSoup
-
 
 protocol EPUBSpreadViewDelegate: AnyObject {
     
@@ -24,16 +18,13 @@ protocol EPUBSpreadViewDelegate: AnyObject {
     
     /// Called when the user tapped on an internal link.
     func spreadView(_ spreadView: EPUBSpreadView, didTapOnInternalLink href: String, tapData: TapData?)
-    
+
     /// Called when the pages visible in the spread changed.
     func spreadViewPagesDidChange(_ spreadView: EPUBSpreadView)
     
     /// Called when the spread view needs to present a view controller.
     func spreadView(_ spreadView: EPUBSpreadView, present viewController: UIViewController)
 
-    /// Called when the spread view receives an unknown JavaScript message.
-    func spreadView(_ spreadView: EPUBSpreadView, userContentController: WKUserContentController, didReceive message: WKScriptMessage)
-    
 }
 
 class EPUBSpreadView: UIView, Loggable, PageView {
@@ -102,14 +93,7 @@ class EPUBSpreadView: UIView, Loggable, PageView {
         registerJSMessages()
 
         NotificationCenter.default.addObserver(self, selector: #selector(voiceOverStatusDidChange), name: Notification.Name(UIAccessibilityVoiceOverStatusChanged), object: nil)
-        
-        UIMenuController.shared.menuItems = [
-            UIMenuItem(
-                title: R2NavigatorLocalizedString("EditingAction.share"),
-                action: #selector(shareSelection)
-            )
-        ]
-        
+
         updateActivityIndicator()
         loadSpread()
     }
@@ -250,18 +234,21 @@ class EPUBSpreadView: UIView, Loggable, PageView {
             let text = selection["text"] as? String,
             let frame = selection["frame"] as? [String: Any] else
         {
+            editingActions.selection = nil
             log(.warning, "Invalid body for selectionDidChange: \(body)")
             return
         }
-        editingActions.selectionDidChange((
-            text: text,
+        editingActions.selection = Selection(
+            locator: Locator(
+                href: "", type: "", text: Locator.Text(highlight: text)
+            ),
             frame: CGRect(
                 x: frame["x"] as? CGFloat ?? 0,
                 y: frame["y"] as? CGFloat ?? 0,
                 width: frame["width"] as? CGFloat ?? 0,
                 height: frame["height"] as? CGFloat ?? 0
             )
-        ))
+        )
     }
     
     /// Called when the user hit the Share item in the selection context menu.
@@ -348,8 +335,7 @@ class EPUBSpreadView: UIView, Loggable, PageView {
     private static let cfiScript = loadScript(named: "epub-cfi")
 
     class func loadScript(named name: String) -> String {
-        return Bundle(for: EPUBSpreadView.self)
-            .url(forResource: "Scripts/\(name)", withExtension: "js")
+        return Bundle.module.url(forResource: "\(name)", withExtension: "js", subdirectory: "Assets/Scripts")
             .flatMap { try? String(contentsOf: $0) }!
     }
     
@@ -438,11 +424,10 @@ extension EPUBSpreadView: WKScriptMessageHandler {
 
     /// Handles incoming calls from JS.
     func userContentController(_ userContentController: WKUserContentController, didReceive message: WKScriptMessage) {
-        if let handler = JSMessages[message.name]  {
-            handler(message.body)
-        } else {
-            delegate?.spreadView(self, userContentController: userContentController, didReceive: message)
+        guard let handler = JSMessages[message.name] else {
+            return
         }
+        handler(message.body)
     }
 
 }
@@ -492,8 +477,6 @@ extension EPUBSpreadView: UIScrollViewDelegate {
 
 extension EPUBSpreadView: WKUIDelegate {
     
-    // The property allowsLinkPreview is default false in iOS9, so it should be safe to use @available(iOS 10.0, *)
-    @available(iOS 10.0, *)
     func webView(_ webView: WKWebView, shouldPreviewElement elementInfo: WKPreviewElementInfo) -> Bool {
         // Preview allowed only if the link is not internal
         return (elementInfo.linkURL?.host != publication.baseURL?.host)